requirements.txt
feedparser
google-genai
google-api-python-client
notion-client
requests
python-dateutil
